23.06.22 | Spring Security | Spring MVC | AppRepair | End of BackEnd

+++ src/main/java/Web/Services/Impl/UserService.java
@@ -64,7 +64,7 @@ public class UserService implements UserDetailsService {
         return false;
     }
 
-    public List<User> usergtList(Long idMin) {
+    public List<User> usergtList(Integer idMin) {
         return manager.createQuery("SELECT u FROM User u WHERE u.id > :paramId", User.class)
                 .setParameter("paramId", idMin).getResultList();
     }
